refactor: migrate date provider and UUID packages to shared module
Moved `undecided.erp.common.dateProvider` and `undecided.erp.common.uuidV7Provider` to `undecided.shared.common`. Updated all related package references and resolved formatting inconsistencies to enhance modularity and reusability.
