--- backend/src/main/java/undecided/erp/common/dateProvider/package-info.java
@@ -1,4 +1,0 @@
-/**
- * デートプロバイダーを格納.
- */
-package undecided.erp.common.dateProvider;
