+++ backend/src/main/java/undecided/erp/common/entity/UuidV7IdValue.java
@@ -4,7 +4,7 @@ import static undecided.shared.common.primitiveOld.Objects2.isNull;
 
 import java.io.Serializable;
 import java.util.UUID;
-import undecided.erp.common.uuidV7Provider.UuidV7Provider;
+import undecided.shared.common.uuidV7Provider.UuidV7Provider;
 
 /**
  * UuidV7IdValueクラスは、UUIDバージョン7を取り扱うための値オブジェクトクラスです。
