+++ backend/src/main/java/undecided/shared/common/dateProvider/StaticDateTimeProvider.java
@@ -1,20 +1,18 @@
-package undecided.erp.common.dateProvider;
+package undecided.shared.common.dateProvider;
 
 import java.time.LocalDateTime;
 import java.util.concurrent.atomic.AtomicReference;
 
 /**
  * StaticDateTimeProvider クラスは、DateProvider 抽象クラスの具体的な実装です。
- * <p>
- * これにより、now() メソッドを呼び出すときに戻される固定の LocalDateTime を設定できます。
+ *
+ * <p>これにより、now() メソッドを呼び出すときに戻される固定の LocalDateTime を設定できます。
  */
 public class StaticDateTimeProvider extends DateProvider {
 
   private static final AtomicReference<LocalDateTime> localDateTime = new AtomicReference<>();
 
-  private StaticDateTimeProvider() {
-
-  }
+  private StaticDateTimeProvider() {}
 
   /**
    * 与えられたLocalDateTimeの値でStaticDateTimeProviderを初期化します。
@@ -25,26 +23,19 @@ public class StaticDateTimeProvider extends DateProvider {
     StaticDateTimeProvider instance = new StaticDateTimeProvider();
     instance.setLocalDateTime(localDateTime);
     new DateProvider(instance);
-
   }
 
-  /**
-   * DateProviderを初期化することでクリアします。
-   */
+  /** DateProviderを初期化することでクリアします。 */
   public static void clear() {
     DateProvider.clear();
-
   }
 
   private void setLocalDateTime(LocalDateTime aLocalDateTime) {
     StaticDateTimeProvider.localDateTime.set(aLocalDateTime);
-
   }
 
   @Override
   protected LocalDateTime now() {
     return localDateTime.get();
-
   }
-
 }
