+++ backend/src/main/java/undecided/shared/common/dateProvider/package-info.java
@@ -1,0 +1,2 @@
+/** デートプロバイダーを格納. */
+package undecided.shared.common.dateProvider;
