+++ backend/src/test/java/undecided/erp/common/entity/UuidV7IdValueTest.java
@@ -7,7 +7,7 @@ import java.util.UUID;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
-import undecided.erp.common.uuidV7Provider.UuidV7Provider;
+import undecided.shared.common.uuidV7Provider.UuidV7Provider;
 
 @DisplayName("UuidV7IdValueのテスト")
 class UuidV7IdValueTest {
